separate p2p plugin from sharp plugin
