+++ src/utils.c
@@ -13,6 +13,17 @@
 #include <ctype.h>
 #include <fcntl.h>
 
+ncclResult_t ncclIbMalloc(void** ptr, size_t size) {
+  size_t page_size = sysconf(_SC_PAGESIZE);
+  void* p;
+  int size_aligned = ROUNDUP(size, page_size);
+  int ret = posix_memalign(&p, page_size, size_aligned);
+  if (ret != 0) return ncclSystemError;
+  memset(p, 0, size);
+  *ptr = p;
+  return ncclSuccess;
+}
+
 int parseStringList(const char* string, struct netIf* ifList, int maxList) {
   if (!string) return 0;
